pyproject.toml
[project]
name = "traffic-arbitration"
version = "0.1.0"
description = ""
authors = [
    {name = "andrey",email = "supervipuser404@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "beautifulsoup4 (>=4.13.1,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false